Added more features to the Material class in API. Need to add features to the database side.
Migrated back to Neo4j because of its awesome lucene index support for
partial string matching in queries. I also created a Spring service to
encapsulate the graph database connection. It is a singleton model and
it also launchs a web admin server on localhost:7474 so you can see the
graph.